requirements.txt
pillow~=12.1.1
datasets~=4.8.4
evaluate~=0.4.6
numpy~=2.4.4
torch~=2.6.0+cu124
transformers~=5.4.0
scikit-learn~=1.8.0
pathlib~=1.0.1
python-dotenv~=1.2.2
pandas~=3.0.1
matplotlib~=3.10.8
ipython~=9.12.0